Move classes to more appropriate packages/places
