+++ src/main/java/org/janelia/saalfeldlab/n5/bdv/metadata/AffineTransform3DJsonAdapter.java
@@ -14,7 +14,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
  */
-package org.janelia.saalfeldlab.n5.bdv;
+package org.janelia.saalfeldlab.n5.bdv.metadata;
 
 import java.lang.reflect.Type;
