+++ src/main/java/org/janelia/saalfeldlab/n5/bdv/metadata/N5Metadata.java
@@ -2,7 +2,6 @@ package org.janelia.saalfeldlab.n5.bdv.metadata;
 
 import com.google.gson.GsonBuilder;
 import net.imglib2.realtransform.AffineTransform3D;
-import org.janelia.saalfeldlab.n5.bdv.AffineTransform3DJsonAdapter;
 
 // NB: marker interface
 public interface N5Metadata {
